asdas

+++ j_com/src/Anal/AUtil.java
@@ -3,8 +3,8 @@ package Anal;
 import basic.Task;
 
 public class AUtil {
-	public static int computeRBF(Task[] ts,int i, int t) {
-		int r=0;
+	public static double computeRBF(Task[] ts,int i, int t) {
+		double r=0;
 		for(int j=0;j<i;j++) {
 			if(j==i-1) {
 				r+=ts[j].exec;
